[MERGE] Fix linux build warnings (#181)
* Fixed spurious module load exception message

* Improve setup.py so it handle SRCDIR issues better

* Fixed bad return code handling in python start('wait') command

* Implemented python validation routine

* Added flag to show output on failure

* More fixed to python setup.py to handle buildling in build directory

* Cleanup of validate.py code

* Bug fix to test and timer output

* Change all mutex state error messages to debug messages

* Added a comment about environ save/restore state when running in parallel

* Added title and version objects to python module

* Added python info to configure.ac

* Fixed to python validation to help ensure the test is correct

* Removed commit path from python module __version__

* Fixed problems with python get() when called using class:id syntax

Also added id to returned data on get_object()

* Added name to keys returned by python get_object()

* Allow classless object find syntax (find by id only)

* Fixed python module loader so errors are reported

* Fixed exceptions not stopping run during event handlers

* Merge of master into dchassin/python

* Fixed missing temporary C linkages for python module

* Working fix for python event return values

* Fixed problem with event handler return values

This fix requires an update to the docs about event handler return values.

* Fixed exception handling in python event handlers

* Update .gitignore

* Fixed build warnings in python module

* Fixed linux warnings in gldcore/{exec,module}.cpp

* Fixed gldcore/property.cpp warnings in linux

* Fix warnings in gldcore/{globals,find}.cpp

* Fixed some linux build warnings

* Fixed extern/cdecl usage to resolve linux warning errors

* Fixed a missing EXPORT for PLC controllers

* Fixed more python warnings

* Fixed warnings in superLU_MT

* Fixed warnings in assert

* Fixed warnings in climate

* Fixed warnings in commercial

* Fixed warnings in generators

* Fixed warnings in market module

* Fixed warnings in powerflow

* Fixed warnings in residential module

* Fixed warnings about PC_NOSYNC in climate and powerflow

* Fixed warnings in tape module

* Fixed warning commercial HVAC class

* Fixed warning in gismo module

* Fixed warnings in mysql module

* Change nan to use stdlib value

* Fixed warnings in CBLAS

* Revert last commit

* Fixed warnings in CBLAS

* Missed one in CBLAS

* Fixed warnings in superLU_MT

* Fixed warnings in superLU_MT

* Fixed warnings in superLU_MT

* Fixed warning in superLU_MT

* Fixed warning in SuperLU_MT

* Fixed warning in superLU_MT

* Fixed warning in gldcore/solver/etp

* Dchassin/docker warnings (#182)

* Fixed warnings in climate
* Fixed warnings in commercial module
* Removed unsupported items as 4.2 from configure.ac
* Fixed warnings in generator module
* Fixed warnings in market module
* Fixed error in climate module
* Fixed warnings in mysql
* Fixed powerflow module warnings
* Fixed warnings in residential module
* Fixed error in climate/csv_reader.cpp
* Fixed warnings in tape module
* Fixed warnings in gldcore

* Fixed warning errors in gldcore and powerflow

* Fixed warning in gldcore/object.cpp

* Fixed warning in python module

* Fixed warnings in gldcore/validate.cpp

* Fixed python build/clean issue and default build flags

* Fixed default build flags

+++ gldcore/class.cpp
@@ -368,7 +368,7 @@ const char *class_get_property_typexsdname(PROPERTYTYPE type) /**< the property 
  **/
 PROPERTYTYPE class_get_propertytype_from_typename(char *name) /**< a string containing the name of the property type */
 {
-	int i;
+	size_t i;
 	for (i=0; i<sizeof(property_type)/sizeof(property_type[0]); i++)
 	{
 		if (strcmp(property_type[i].name,name)==0)
@@ -1351,7 +1351,7 @@ int class_get_xsd(CLASS *oclass, /**< a pointer to the class to convert to XSD *
 {
 	size_t n=0;
 	PROPERTY *prop;
-	int i;
+	size_t i;
 	CLASS *oc = oclass;
 	extern KEYWORD oflags[];
 	struct {
@@ -1374,7 +1374,7 @@ int class_get_xsd(CLASS *oclass, /**< a pointer to the class to convert to XSD *
 	n += buffer_write(buffer+n, len-n, "<xs:element name=\"%s\">\n", oclass->name);
 	n += buffer_write(buffer+n, len-n, "\t<xs:complexType>\n");
 	n += buffer_write(buffer+n, len-n, "\t\t<xs:all>\n");
-	for (i=0; i < sizeof(attribute) / sizeof(attribute[0]); i++)
+	for ( i = 0 ; i < sizeof(attribute) / sizeof(attribute[0]) ; i++ )
 	{
 		n += buffer_write(buffer+n, len-n, "\t\t\t<xs:element name=\"%s\">\n", attribute[i].name);
 		n += buffer_write(buffer+n, len-n, "\t\t\t\t<xs:simpleType>\n");
